refactor(menu): create LabelRepository
