+++ menu/src/main/java/de/hpi/android/menu/data/LabelDto.kt
@@ -1,0 +1,8 @@
+package de.hpi.android.menu.data
+
+import de.hpi.android.core.data.Dto
+
+data class LabelDto(
+    override val id: String,
+    val title: String
+): Dto<LabelDto>
